requirements.txt
fastapi
uvicorn
langchain
langchain-community
python-dotenv